test: Add tests for meta mutators
