+++ pkg/handlers/docker/mutation/customimage/inject_test.go
@@ -6,142 +6,16 @@ package customimage
 import (
 	"testing"
 
-	"github.com/onsi/gomega"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/types"
-	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
-	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
-
 	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/capi/clustertopology/handlers/mutation"
-	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/testutils/capitest"
-	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/testutils/capitest/request"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/docker/mutation/customimage/tests"
 )
 
-func newWorkerDockerMachineTemplateRequestItem(
-	uid types.UID,
-) runtimehooksv1.GeneratePatchesRequestItem {
-	return request.NewRequestItem(
-		&unstructured.Unstructured{
-			Object: map[string]interface{}{
-				"apiVersion": "infrastructure.cluster.x-k8s.io/v1beta1",
-				"kind":       "DockerMachineTemplate",
-				"metadata": map[string]interface{}{
-					"name":      "docker-machine-template",
-					"namespace": "docker-cluster",
-				},
-			},
-		},
-		&runtimehooksv1.HolderReference{
-			APIVersion: clusterv1.GroupVersion.String(),
-			Kind:       "MachineDeployment",
-			FieldPath:  "spec.template.spec.infrastructureRef",
-		},
-		uid,
-	)
-}
-
-func newCPDockerMachineTemplateRequestItem(
-	uid types.UID,
-) runtimehooksv1.GeneratePatchesRequestItem {
-	return request.NewRequestItem(
-		&unstructured.Unstructured{
-			Object: map[string]interface{}{
-				"apiVersion": "infrastructure.cluster.x-k8s.io/v1beta1",
-				"kind":       "DockerMachineTemplate",
-				"metadata": map[string]interface{}{
-					"name":      "docker-machine-template",
-					"namespace": "docker-cluster",
-				},
-			},
-		},
-		&runtimehooksv1.HolderReference{
-			APIVersion: clusterv1.GroupVersion.String(),
-			FieldPath:  "spec.machineTemplate.infrastructureRef",
-		},
-		uid,
-	)
-}
-
 func TestGeneratePatches(t *testing.T) {
-	capitest.ValidateGeneratePatches(
+	t.Parallel()
+
+	tests.TestGeneratePatches(
 		t,
 		func() mutation.GeneratePatches { return NewPatch() },
-		capitest.PatchTestDef{
-			Name: "image unset for workers",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					"builtin",
-					apiextensionsv1.JSON{
-						Raw: []byte(
-							`{"machineDeployment": {"class": "a-worker", "version": "v1.2.3"}}`,
-						),
-					},
-				),
-			},
-			RequestItem: newWorkerDockerMachineTemplateRequestItem("1234"),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/customImage",
-				ValueMatcher: gomega.Equal("ghcr.io/mesosphere/kind-node:v1.2.3"),
-			}},
-		},
-		capitest.PatchTestDef{
-			Name: "image set for workers",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					variableName,
-					"a-specific-image",
-				),
-				capitest.VariableWithValue(
-					"builtin",
-					apiextensionsv1.JSON{
-						Raw: []byte(`{"machineDeployment": {"class": "a-worker"}}`),
-					},
-				),
-			},
-			RequestItem: newWorkerDockerMachineTemplateRequestItem("1234"),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/customImage",
-				ValueMatcher: gomega.Equal("a-specific-image"),
-			}},
-		},
-		capitest.PatchTestDef{
-			Name: "image unset for control plane",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					"builtin",
-					apiextensionsv1.JSON{Raw: []byte(`{"controlPlane": {"version": "v1.2.3"}}`)},
-				),
-			},
-			RequestItem: newCPDockerMachineTemplateRequestItem("1234"),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/customImage",
-				ValueMatcher: gomega.Equal("ghcr.io/mesosphere/kind-node:v1.2.3"),
-			}},
-		},
-		capitest.PatchTestDef{
-			Name: "image set for control plane",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					variableName,
-					"a-specific-image",
-				),
-				capitest.VariableWithValue(
-					"builtin",
-					apiextensionsv1.JSON{
-						Raw: []byte(`{"machineDeployment": {"class": "a-worker"}}`),
-					},
-				),
-			},
-			RequestItem: newCPDockerMachineTemplateRequestItem("1234"),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/customImage",
-				ValueMatcher: gomega.Equal("a-specific-image"),
-			}},
-		},
+		VariableName,
 	)
 }
