+++ pkg/handlers/generic/mutation/etcd/variables.go
@@ -20,8 +20,8 @@ var (
 )
 
 const (
-	// variableName is the external patch variable name.
-	variableName = "etcd"
+	// VariableName is the external patch variable name.
+	VariableName = "etcd"
 
 	// HandlerNameVariable is the name of the variable handler.
 	HandlerNameVariable = "etcdVars"
@@ -43,7 +43,7 @@ func (h *etcdVariableHandler) DiscoverVariables(
 	resp *runtimehooksv1.DiscoverVariablesResponse,
 ) {
 	resp.Variables = append(resp.Variables, clusterv1.ClusterClassVariable{
-		Name:     variableName,
+		Name:     VariableName,
 		Required: false,
 		Schema:   v1alpha1.Etcd{}.VariableSchema(),
 	})
