+++ pkg/handlers/generic/mutation/etcd/variables_test.go
@@ -15,7 +15,7 @@ import (
 func TestVariableValidation(t *testing.T) {
 	capitest.ValidateDiscoverVariables(
 		t,
-		variableName,
+		VariableName,
 		ptr.To(v1alpha1.Etcd{}.VariableSchema()),
 		false,
 		NewVariable,
