+++ pkg/handlers/generic/mutation/extraapiservercertsans/inject_test.go
@@ -6,42 +6,16 @@ package extraapiservercertsans
 import (
 	"testing"
 
-	. "github.com/onsi/gomega"
-	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
-
-	"github.com/d2iq-labs/capi-runtime-extensions/api/v1alpha1"
 	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/capi/clustertopology/handlers/mutation"
-	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/testutils/capitest"
-	"github.com/d2iq-labs/capi-runtime-extensions/common/pkg/testutils/capitest/request"
+	"github.com/d2iq-labs/capi-runtime-extensions/pkg/handlers/generic/mutation/extraapiservercertsans/tests"
 )
 
 func TestGeneratePatches(t *testing.T) {
-	capitest.ValidateGeneratePatches(
+	t.Parallel()
+
+	tests.TestGeneratePatches(
 		t,
 		func() mutation.GeneratePatches { return NewPatch() },
-		capitest.PatchTestDef{
-			Name: "unset variable",
-		},
-		capitest.PatchTestDef{
-			Name: "extra API server cert SANs set",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					variableName,
-					v1alpha1.ExtraAPIServerCertSANs{"a.b.c.example.com", "d.e.f.example.com"},
-				),
-			},
-			RequestItem: request.NewKubeadmControlPlaneTemplateRequestItem(""),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation: "add",
-				Path:      "/spec/template/spec/kubeadmConfigSpec/clusterConfiguration",
-				ValueMatcher: HaveKeyWithValue(
-					"apiServer",
-					HaveKeyWithValue(
-						"certSANs",
-						[]interface{}{"a.b.c.example.com", "d.e.f.example.com"},
-					),
-				),
-			}},
-		},
+		VariableName,
 	)
 }
