+++ pkg/handlers/generic/mutation/extraapiservercertsans/variables.go
@@ -20,8 +20,8 @@ var (
 )
 
 const (
-	// variableName is the external patch variable name.
-	variableName = "extraAPIServerCertSANs"
+	// VariableName is the external patch variable name.
+	VariableName = "extraAPIServerCertSANs"
 
 	// HandlerNameVariable is the name of the variable handler.
 	HandlerNameVariable = "ExtraAPIServerCertSANsVars"
@@ -43,7 +43,7 @@ func (h *extraAPIServerCertSANsVariableHandler) DiscoverVariables(
 	resp *runtimehooksv1.DiscoverVariablesResponse,
 ) {
 	resp.Variables = append(resp.Variables, clusterv1.ClusterClassVariable{
-		Name:     variableName,
+		Name:     VariableName,
 		Required: false,
 		Schema:   v1alpha1.ExtraAPIServerCertSANs{}.VariableSchema(),
 	})
