+++ pkg/handlers/generic/mutation/httpproxy/variables.go
@@ -20,8 +20,8 @@ var (
 )
 
 const (
-	// variableName is the external patch variable name.
-	variableName = "proxy"
+	// VariableName is the external patch variable name.
+	VariableName = "proxy"
 
 	// HandlerNameVariable is the name of the variable handler.
 	HandlerNameVariable = "HTTPProxyVars"
@@ -43,7 +43,7 @@ func (h *httpProxyVariableHandler) DiscoverVariables(
 	resp *runtimehooksv1.DiscoverVariablesResponse,
 ) {
 	resp.Variables = append(resp.Variables, clusterv1.ClusterClassVariable{
-		Name:     variableName,
+		Name:     VariableName,
 		Required: false,
 		Schema:   v1alpha1.HTTPProxy{}.VariableSchema(),
 	})
