+++ pkg/handlers/generic/mutation/httpproxy/variables_test.go
@@ -15,7 +15,7 @@ import (
 func TestVariableValidation(t *testing.T) {
 	capitest.ValidateDiscoverVariables(
 		t,
-		variableName,
+		VariableName,
 		ptr.To(v1alpha1.HTTPProxy{}.VariableSchema()),
 		false,
 		NewVariable,
